Extracted ImageComparison class
